requirements.txt
python-telegram-bot>=21.0
aiohttp
mutagen
python-dotenv
